requirements.txt


openai == 1.43.0
python-jobspy == 1.1.65

